% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_ei_data.R
\name{analyze_ei_data}
\alias{analyze_ei_data}
\title{Evaluate NM, ER, MOB, and monotone EI bounds (with optional confidence intervals)
to be summarized in plot() or summarize()}
\usage{
analyze_ei_data(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An EI data object, created with make_ei_data()}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\value{
A list of all methods and their estimates and bootstrap draws.
}
\description{
Evaluate NM, ER, MOB, and monotone EI bounds (with optional confidence intervals)
to be summarized in plot() or summarize()
}
