% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ei_data.R
\name{is_ei_data}
\alias{is_ei_data}
\title{Test if an EI data object is an EI data object}
\usage{
is_ei_data(data)
}
\arguments{
\item{data}{A potential EI data object}
}
\value{
True if an EI data object was passed, False otherwise
}
\description{
Test if an EI data object is an EI data object
}
