% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ei_data.R
\name{make_ei_data}
\alias{make_ei_data}
\title{Create an EI data object}
\usage{
make_ei_data(data = NA, Xn = NA, Yn = NA, wt = NA)
}
\arguments{
\item{data}{A data.frame with columns for Xn, Yn (and weight if desired)}

\item{Xn}{A vector for Xn}

\item{Yn}{A vector for Yn}

\item{wt}{A vector for weight, if desired}
}
\value{
An EI data object with features Xn and Yn (and weight if applicable)
}
\description{
Create an EI data object
}
