% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOB.R
\name{MOB}
\alias{MOB}
\title{Calculate method of bounds}
\usage{
MOB(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An EI data object}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\value{
A list of: (MOB_1: (MOB lower bound for group 1, MOB upper bound for group 1),
                    MOB_0: (MOB lower bound for group 0, MOB upper bound for group 0),
                    MOB_disparity: (MOB lower bound for disparity, MOB upper bound for disparity),
                    MOB_1_conf: (MOB confidence interval lower bound for group 1, MOB confidence interval upper bound for group 1),
                    MOB_0_conf: (MOB confidence interval lower bound for group 0, MOB confidence interval upper bound for group 0),
                    MOB_disparity_conf: (MOB confidence interval lower bound for disparity, MOB confidence interval upper bound for disparity))
}
\description{
Calculate method of bounds
}
