% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motone_EI.R
\name{monotone_EI}
\alias{monotone_EI}
\title{Calculate monotone ecological inference bounds based on the neighborhood model,
ecological regression, and the method of bounds.}
\usage{
monotone_EI(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An EI data object}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\value{
A list of: (monotone_1: (lower bound for group 1, upper bound for group 1),
                    monotone_0: (lower bound for group 0, upper bound for group 0),
                    monotone_disparity: (lower bound for disparity, upper bound for disparity),
                    monotone_1_conf: (lower 90\% confidence interval bound for group 1, upper 90\% confidence interval bound for group 1),
                    monotone_0_conf: (lower 90\% confidence interval bound for group 0, upper 90\% confidence interval bound for group 0),
                    monotone_disparity_conf: (lower 90\% confidence interval bound for disparity, upper 90\% confidence interval bound for disparity))
}
\description{
Calculate monotone ecological inference bounds based on the neighborhood model,
ecological regression, and the method of bounds.
}
