% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NM.R
\name{NM}
\alias{NM}
\title{Calculate neighborhood model estimate}
\usage{
NM(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An EI data object}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\value{
A list of: (NM_1: Neighborhood Model estimate for group 1,
                    NM_0: Neighborhood Model estimate for group 0,
                    NM_disparity: Neighborhood Model disparity estimate,
                    NM_1_conf: 90\% Confidence interval around group 1,
                    NM_0_conf: 90\% Confidence interval around group 0,
                    NM_disparity_conf: 90\% Confidence interval around disparity)
}
\description{
Calculate neighborhood model estimate
}
