% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.}
\usage{
plot(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An analyzed EI data object from analyze_ei_data()}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\description{
Plot monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.
}
