% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.}
\usage{
summarize(data)
}
\arguments{
\item{data}{An analyzed EI data object from analyze_ei_data()}
}
\description{
Summarize monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.
}
