% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_vax.R
\docType{data}
\name{voter_vax}
\alias{voter_vax}
\title{Proportion of Republican votes in the 2020 Presidential election and COVID-19 vaccination rate by county}
\format{
## `voter_vax`
A data frame with 4532 rows and 4 columns
\describe{
  \item{county_fips}{County FIPS code}
  \item{prop_republican}{Proportion Republican votes in the 2020 presidential election in the county}
  \item{total_votes}{Total votes cast in the 2020 presidential election in the county}
  \item{prop_covid_vax}{Proportion of county fully vaccinated for COVID as of 12/31/2021}
  ...
}
}
\source{
<https://data.cdc.gov/Vaccinations/COVID-19-Vaccinations-in-the-United-States-County/8xkx-amqh/data_preview>
   <https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/VOQCHQ>
}
\usage{
voter_vax
}
\description{
Proportion of Republican votes in the 2020 Presidential election and COVID-19 vaccination rate by county
}
\keyword{datasets}
