% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.}
\usage{
summarize(data, n_bootstrap_replicates = 0)
}
\arguments{
\item{data}{An EI data object}

\item{n_bootstrap_replicates}{The optional number of bootstrap replicates to obtain 95\% confidence intervals}
}
\description{
Summarize monotone ecological inference, neighborhood model, ecological
regression, and method of bounds by group and for the disparity.
}
